Move Inventory.Domain.Item
Move module to its own file.
